backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiverse-copilot-backend"
version = "0.1.0"
description = "Backend for Multiverse Copilot"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pydantic>=2.8.0",
  "pydantic-settings>=2.4.0",
  "boto3>=1.35.0",
  "python-multipart>=0.0.9",
  "pypdf>=4.3.0",
  "numpy>=1.26.0",
  "faiss-cpu>=1.8.0",
  "tenacity>=8.5.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

